Revert "[Artifacts] Add the option to calculate the size of artifact (#4719)"
This reverts commit db3f1874dcdc5f6b42508396af3d3e9fd24f2283.
