+++ server/api/api/endpoints/artifacts.py
@@ -88,7 +88,6 @@ async def store_artifact(
         iter=iter,
         project=project,
         producer_id=tree,
-        auth_info=auth_info,
     )
     return {}
