+++ server/api/api/endpoints/artifacts_v2.py
@@ -125,7 +125,6 @@ async def store_artifact(
         iter,
         project,
         producer_id=producer_id,
-        auth_info=auth_info,
     )
     return await run_in_threadpool(
         server.api.crud.Artifacts().get_artifact,
