+++ tests/integration/sdk_api/base.py
@@ -42,13 +42,13 @@ class TestMLRunIntegration:
     db_name = "mlrun"
     db_dsn = f"mysql+pymysql://{db_user}@{db_host_internal}:{db_port}/{db_name}"
 
-    def setup_method(self, method, extra_env=None):
+    def setup_method(self, method):
         self._logger = logger
         self._logger.info(
             f"Setting up test {self.__class__.__name__}::{method.__name__}"
         )
         self._run_db()
-        api_url = self._run_api(extra_env)
+        api_url = self._run_api()
         self._test_env = {}
         self._old_env = {}
         self._setup_env({"MLRUN_DBPATH": api_url})
@@ -131,7 +131,7 @@ class TestMLRunIntegration:
 
         self._ensure_database_liveness(timeout=self.db_liveness_timeout)
 
-    def _run_api(self, extra_env=None):
+    def _run_api(self):
         self._logger.debug("Starting API")
         self._run_command(
             # already compiled schemas in run-test-db
@@ -143,8 +143,7 @@ class TestMLRunIntegration:
                     "MLRUN_HTTPDB__DSN": self.db_dsn,
                     "MLRUN_LOG_LEVEL": "DEBUG",
                     "MLRUN_SECRET_STORES__TEST_MODE_MOCK_SECRETS": "True",
-                },
-                extra_env,
+                }
             ),
             cwd=TestMLRunIntegration.root_path,
         )
@@ -207,11 +206,9 @@ class TestMLRunIntegration:
         self._logger.debug("Database ready for connection")
 
     @staticmethod
-    def _extend_current_env(env, extra_env=None):
+    def _extend_current_env(env):
         current_env = copy.deepcopy(os.environ)
         current_env.update(env)
-        if extra_env:
-            current_env.update(extra_env)
         return current_env
 
     @staticmethod
